refactor email rendering and sending
Add a new function, satchmo_store.mail.send_store_email(). Code that
want to send email can use this to do things like the following easily:

  - format the subject to use Config.store_name
  - render message body with template/context
  - handle mail errors

As a side-effect, all users of email use the settings.ADMINS as a
fallback email in the situation that Config.store_email is not defined.